Add files via upload
A beautiful Frontend website for your dogs that is build using flask Bootstrap Html Css. you only need to download the code and insert the require files in required Directory like the all the images need to  be in static file then in images folder and the css file need to be inside the static/assets/css and the index file templates/index.html.